Moved deepCopy, deepAssign, deepCopyAdded, deepAssignAdded to a new file, deep.js
